refactor: update imports to use v3.0 packages in v3.1-RC2 files and remove v3.1-RC1
