--- models/golang/v3.1-RC/gbfs/gbfs.go
@@ -1,26 +1,0 @@
-// Copyright 2024 MobilityData
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package gbfs
-
-import (
-	gbfs_v30 "github.com/MobilityData/gbfs-json-schema/models/golang/v3.0/gbfs"
-)
-
-// Auto-discovery file that links to all of the other files published by the system.
-type Gbfs struct {
-	gbfs_v30.Gbfs
-}
-
-
