--- models/golang/v3.1-RC/system_regions/system_regions.go
@@ -1,24 +1,0 @@
-// Copyright 2024 MobilityData
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package system_regions
-
-import (
-	system_regions_v30 "github.com/MobilityData/gbfs-json-schema/models/golang/v3.0/system_regions"
-)
-
-// Describes regions for a system that is broken up by geographic or political region.
-type SystemRegions struct {
-	system_regions_v30.SystemRegions 
-}
