+++ models/golang/v3.1-RC2/station_information/station_information.go
@@ -15,13 +15,12 @@
 package station_information
 
 import (
-	station_information_v31RC "github.com/MobilityData/gbfs-json-schema/models/golang/v3.1-RC/station_information"
 	station_information_v30 "github.com/MobilityData/gbfs-json-schema/models/golang/v3.0/station_information"
 )
 
 // List of all stations, their capacities and locations. REQUIRED of systems utilizing docks.
 type StationInformation struct {
-	station_information_v31RC.StationInformation
+	station_information_v30.StationInformation
 	// Array that contains one object per station as defined below.
 	Data Data `json:"data"`
 }
