+++ models/golang/v3.1-RC2/vehicle_types/vehicle_types.go
@@ -15,11 +15,11 @@
 package vehicle_types
 
 import (
-	vehicle_types_v31RC "github.com/MobilityData/gbfs-json-schema/models/golang/v3.1-RC/vehicle_types"
+	vehicle_types_v30 "github.com/MobilityData/gbfs-json-schema/models/golang/v3.0/vehicle_types"
 )
 
 // Describes the types of vehicles that System operator has available for rent (added in
 // v2.1-RC).
 type VehicleTypes struct {
-	vehicle_types_v31RC.VehicleTypes
+	vehicle_types_v30.VehicleTypes
 }
